+++ base/funciones/migra.f__on_mig_tpm_moneda_tmoneda.sql
@@ -1,8 +1,22 @@
-CREATE OR REPLACE FUNCTION migra.f__on_trig_tpm_moneda_tmoneda (
-						  v_operacion varchar,p_id_moneda int4,p_codigo varchar,p_estado_reg varchar,p_fecha_mod timestamp,p_fecha_reg timestamp,p_id_usuario_mod int4,p_id_usuario_reg int4,p_moneda varchar,p_origen varchar,p_prioridad int4,p_tipo_actualizacion varchar,p_tipo_moneda varchar)
-						RETURNS text AS
-						$BODY$
+--------------- SQL ---------------
 
+CREATE OR REPLACE FUNCTION migra.f__on_trig_tpm_moneda_tmoneda (
+  v_operacion varchar,
+  p_id_moneda integer,
+  p_codigo varchar,
+  p_estado_reg varchar,
+  p_fecha_mod timestamp,
+  p_fecha_reg timestamp,
+  p_id_usuario_mod integer,
+  p_id_usuario_reg integer,
+  p_moneda varchar,
+  p_origen varchar,
+  p_prioridad integer,
+  p_tipo_actualizacion varchar,
+  p_tipo_moneda varchar
+)
+RETURNS text AS
+$body$
 /*
 						Function:  Para migracion de la tabla param.tgestion
 						Fecha Creacion:  February 5, 2013, 7:49 am
@@ -46,31 +60,52 @@ CREATE OR REPLACE FUNCTION migra.f__on_trig_tpm_moneda_tmoneda (
 
 						       
 							    ELSEIF  v_operacion = 'UPDATE' THEN
-						               UPDATE 
+						               
+                                       
+                                IF  not EXISTS(select 1 
+                                           from PARAM.tmoneda 
+                                           where id_moneda=p_id_moneda) THEN
+                                       
+                                            raise exception 'No existe el registro quw  desea modificar';
+                                            
+                                END IF;
+                                       
+                                       
+                                       UPDATE 
 						                  PARAM.tmoneda  
-						                SET						 codigo=p_codigo
-						 ,estado_reg=p_estado_reg
-						 ,fecha_mod=p_fecha_mod
-						 ,fecha_reg=p_fecha_reg
-						 ,id_usuario_mod=p_id_usuario_mod
-						 ,id_usuario_reg=p_id_usuario_reg
-						 ,moneda=p_moneda
-						 ,origen=p_origen
-						 ,prioridad=p_prioridad
-						 ,tipo_actualizacion=p_tipo_actualizacion
-						 ,tipo_moneda=p_tipo_moneda
-						 WHERE id_moneda=p_id_moneda;
+						                SET					
+                                         codigo=p_codigo
+                                        ,estado_reg=p_estado_reg
+                                         ,fecha_mod=p_fecha_mod
+                                         ,fecha_reg=p_fecha_reg
+                                         ,id_usuario_mod=p_id_usuario_mod
+                                         ,id_usuario_reg=p_id_usuario_reg
+                                         ,moneda=p_moneda
+                                         ,origen=p_origen
+                                         ,prioridad=p_prioridad
+                                         ,tipo_actualizacion=p_tipo_actualizacion
+                                         ,tipo_moneda=p_tipo_moneda
+                                         WHERE id_moneda=p_id_moneda;
 
 						       
-						       ELSEIF  v_operacion = 'DELETE' THEN
+						 ELSEIF  v_operacion = 'DELETE' THEN
 						       
-						         DELETE FROM 
+						         IF  not EXISTS(select 1 
+                                           from PARAM.tmoneda 
+                                           where id_moneda=p_id_moneda) THEN
+                                       
+                                            raise exception 'No existe el registro quw  desea modificar';
+                                            
+                                 END IF;
+                                 
+                                 
+                                 DELETE FROM 
 						              PARAM.tmoneda
  
 						              						 WHERE id_moneda=p_id_moneda;
 
 						       
-						       END IF;  
+						END IF;  
 						  
 						 return 'true';
 						
@@ -79,11 +114,9 @@ CREATE OR REPLACE FUNCTION migra.f__on_trig_tpm_moneda_tmoneda (
 						--WHEN exception_name THEN
 						--  statements;
 						END;
-						$BODY$
-
-
-						LANGUAGE 'plpgsql'
-						VOLATILE
-						CALLED ON NULL INPUT
-						SECURITY INVOKER
-						COST 100;
+$body$
+LANGUAGE 'plpgsql'
+VOLATILE
+CALLED ON NULL INPUT
+SECURITY INVOKER
+COST 100;
