+++ base/funciones/migra.f__on_mig_tpm_programa_tprograma.sql
@@ -1,8 +1,20 @@
-CREATE OR REPLACE FUNCTION migra.f__on_trig_tpm_programa_tprograma (
-						  v_operacion varchar,p_codigo_programa varchar,p_id_programa int4,p_descripcion_programa text,p_estado_reg varchar,p_fecha_mod timestamp,p_fecha_reg timestamp,p_id_programa_actif int4,p_id_usuario_mod int4,p_id_usuario_reg int4,p_nombre_programa varchar)
-						RETURNS text AS
-						$BODY$
+--------------- SQL ---------------
 
+CREATE OR REPLACE FUNCTION migra.f__on_trig_tpm_programa_tprograma (
+  v_operacion varchar,
+  p_codigo_programa varchar,
+  p_id_programa integer,
+  p_descripcion_programa text,
+  p_estado_reg varchar,
+  p_fecha_mod timestamp,
+  p_fecha_reg timestamp,
+  p_id_programa_actif integer,
+  p_id_usuario_mod integer,
+  p_id_usuario_reg integer,
+  p_nombre_programa varchar
+)
+RETURNS text AS
+$body$
 /*
 						Function:  Para migracion de la tabla param.tgestion
 						Fecha Creacion:  February 5, 2013, 12:21 pm
@@ -42,7 +54,18 @@ CREATE OR REPLACE FUNCTION migra.f__on_trig_tpm_programa_tprograma (
 
 						       
 							    ELSEIF  v_operacion = 'UPDATE' THEN
-						               UPDATE 
+						              
+                              --chequear si ya existe el auxiliar si no sacar un error
+                                 IF  not EXISTS(select 1 
+                                   from  PARAM.tprograma 
+                                   where id_programa=p_id_programa) THEN
+                                             
+                                    raise exception 'No existe el registro que desea modificar';
+                                                  
+                                 END IF; 
+                                
+                                
+                                 UPDATE 
 						                  PARAM.tprograma  
 						                SET						 codigo_programa=p_codigo_programa
 						 ,descripcion_programa=p_descripcion_programa
@@ -58,7 +81,17 @@ CREATE OR REPLACE FUNCTION migra.f__on_trig_tpm_programa_tprograma (
 						       
 						       ELSEIF  v_operacion = 'DELETE' THEN
 						       
-						         DELETE FROM 
+						        --chequear si ya existe el auxiliar si no sacar un error
+                                 IF  not EXISTS(select 1 
+                                   from  PARAM.tprograma 
+                                   where id_programa=p_id_programa) THEN
+                                             
+                                    raise exception 'No existe el registro que desea eliminar';
+                                                  
+                                 END IF;
+                               
+                               
+                                DELETE FROM 
 						              PARAM.tprograma
  
 						              						 WHERE id_programa=p_id_programa;
@@ -73,11 +106,9 @@ CREATE OR REPLACE FUNCTION migra.f__on_trig_tpm_programa_tprograma (
 						--WHEN exception_name THEN
 						--  statements;
 						END;
-						$BODY$
-
-
-						LANGUAGE 'plpgsql'
-						VOLATILE
-						CALLED ON NULL INPUT
-						SECURITY INVOKER
-						COST 100;
+$body$
+LANGUAGE 'plpgsql'
+VOLATILE
+CALLED ON NULL INPUT
+SECURITY INVOKER
+COST 100;
